:adhesive_bandage: - Switching to `serde_json` from `flexbuffers`. because of a problem with `flexbuffers` not being able to process `BTreeMap<i32, _>`.

+++ kernel/src/io/events/categories.rs
@@ -6,9 +6,9 @@ use std::collections::BTreeMap;
 /// This event that is produced when a [`CategoriesCommand`](crate::io::commands::categories::CategoriesCommand)
 /// is applied to a [`Categories`](crate::entities::categories::Categories) entity.
 #[derive(Debug, Clone, Event, Deserialize, Serialize)]
-#[persist(enc = "flexbuffers::to_vec", dec = "flexbuffers::from_slice")]
+#[persist(enc = "serde_json::to_vec", dec = "serde_json::from_slice")]
 pub enum CategoriesEvent {
     AddedCategory { id: CategoryId },
     RemovedCategory { id: CategoryId },
-    ChangedOrdering { new: BTreeMap<i32, CategoryId> },
+    ChangedOrdering { new: BTreeMap<i64, CategoryId> },
 }
